Addressed latest CRAN checks with _STRICT_R_HEADERS (2024-09-23)
Compilation fails with _R_USE_STRICT_R_HEADERS_=true, which defines
STRICT_R_HEADERS to 1 which removes

- the legacy definition of PI (use POSIX's M_PI, available in R fer ever).
- the RS.h declarations for Calloc, Realloc, Free (use R_ forms i
  available since R 3.4.0).

The aim is to clean the namespace: in particular having a definition
for Free has conflicted with some packages' C++ code.

It is planned that STRICT_R_HEADERS=1 will become the default for 4.5.0,
for which it would be good if all CRAN packages compile/install ok with
the new default.

Your package is among the ones which need updating.  We would thus
really appreciate if you could provide a new version of your package as
soon as possible which checks ok with STRICT_R_HEADERS=1.

You can verify that your package checks ok with STRICT_R_HEADERS=1 via R
CMD check --as-cran using a current version of R-devel.

Please correct before 2024-10-21 to safely retain your package on CRAN.

Updated GitHub workflows